updated assignmentlist.h
I commented out the boolean in the add assignment function - it isn't necessary, since the item can be added to the end of the list once it's gone through the loop. It will only go through the loop if it wasn't already added.

Added an output commented to let the user know when an assignment was completed.
Added the empty conditions for printing the lists, and changed the name to "display" to make it more clear
Added return statements to the display and adding from file functions.